move schema maker helper into aux
