+++ src/Auxillary/CMakeLists.txt
@@ -6,3 +6,4 @@ add_subdirectory(Misc)
 add_subdirectory(Input_Output)
 add_subdirectory(aux_json)
 add_subdirectory(unit_conversion)
+add_subdirectory(json_schema)
