+++ src/Auxillary/json_schema/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+add_library(aux_make_schema STATIC make_schema.cpp)
+
+target_link_libraries(aux_make_schema PRIVATE nlohmann_json)
+
+target_include_directories(aux_make_schema PUBLIC include)
