+++ src/Auxillary/json_schema/include/make_schema.hpp
@@ -1,0 +1,5 @@
+#include <nlohmann/json.hpp>
+
+namespace Aux::schema {
+  nlohmann::json make_list_schema_of(nlohmann::json const &element_schema);
+}
