+++ src/Model/Componentfactory/Componentfactory.cpp
@@ -1,4 +1,5 @@
 #include <Componentfactory.hpp>
+#include <make_schema.hpp>
 
 namespace Model::Componentfactory {
 
@@ -14,15 +15,6 @@ namespace Model::Componentfactory {
   }
   void Componentfactory::add_edge_type(std::unique_ptr<AbstractEdgeType> edgeType) {
     this->edge_type_map.insert({edgeType->get_name(), std::move(edgeType)});
-  }
-
-  nlohmann::json make_list_schema_of(nlohmann::json const &element_schema);
-  nlohmann::json make_list_schema_of(nlohmann::json const &element_schema) {
-    nlohmann::json list_schema;
-    list_schema["type"] = "array";
-    list_schema["items"] = element_schema;
-
-    return list_schema;
   }
 
   nlohmann::json Componentfactory::get_topology_schema() {
@@ -41,11 +33,11 @@ namespace Model::Componentfactory {
 
     auto & node_schemas = topology_schema["properties"]["nodes"]["properties"];
     for (auto const &[name, component] : this->node_type_map) {
-      node_schemas[name] = make_list_schema_of(component->get_schema());
+      node_schemas[name] = Aux::schema::make_list_schema_of(component->get_schema());
     }
     auto & edge_schemas = topology_schema["properties"]["connections"]["properties"];
     for (auto const &[name, component] : this->edge_type_map) {
-      edge_schemas[name] = make_list_schema_of(component->get_schema());
+      edge_schemas[name] = Aux::schema::make_list_schema_of(component->get_schema());
     }
 
     return topology_schema;
